Extract shared utility functions into utils/helpers.ts
Deduplicate downloadFile (3 copies), formatBytes (2 copies),
getFileExtension (2 copies), nowIso (2 copies), and
TranscriptListItem interface (2 copies) into a single shared
module. Export isTerminalStatus from DashboardContext for reuse
in jobs page. Remove getFilenameExtension (replaced with
getFileExtension || 'bin' at the single call site).
