+++ frontend-next/src/utils/helpers.ts
@@ -1,0 +1,63 @@
+export function nowIso(): string {
+  return new Date().toISOString()
+}
+
+export function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => {
+    window.setTimeout(resolve, ms)
+  })
+}
+
+export function getFileExtension(filename: string): string {
+  const dot = filename.lastIndexOf('.')
+  if (dot === -1) return ''
+  return filename.slice(dot + 1).toLowerCase()
+}
+
+export function formatBytes(bytes: number): string {
+  if (!Number.isFinite(bytes) || bytes <= 0) return '0 B'
+  if (bytes < 1024) return `${bytes} B`
+  if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`
+  if (bytes < 1024 * 1024 * 1024) return `${(bytes / (1024 * 1024)).toFixed(1)} MB`
+  return `${(bytes / (1024 * 1024 * 1024)).toFixed(1)} GB`
+}
+
+export function downloadFileBlob(file: File): void {
+  const url = URL.createObjectURL(file)
+  const link = document.createElement('a')
+  link.href = url
+  link.download = file.name
+  document.body.appendChild(link)
+  link.click()
+  document.body.removeChild(link)
+  URL.revokeObjectURL(url)
+}
+
+export function downloadBase64(base64Data: string, filename: string, mimeType: string): void {
+  const byteCharacters = atob(base64Data)
+  const byteNumbers = new Array(byteCharacters.length)
+  for (let i = 0; i < byteCharacters.length; i++) {
+    byteNumbers[i] = byteCharacters.charCodeAt(i)
+  }
+  const byteArray = new Uint8Array(byteNumbers)
+  const blob = new Blob([byteArray], { type: mimeType })
+
+  const link = document.createElement('a')
+  const blobUrl = URL.createObjectURL(blob)
+  link.href = blobUrl
+  link.download = filename
+  document.body.appendChild(link)
+  link.click()
+  setTimeout(() => {
+    document.body.removeChild(link)
+    URL.revokeObjectURL(blobUrl)
+  }, 1500)
+}
+
+export interface TranscriptListItem {
+  media_key: string
+  title_label: string
+  updated_at?: string | null
+  line_count?: number
+  expires_at?: string | null
+}
